Implement base class for natural corruptions
